phc-requirements.txt
google-cloud-storage
fsspec
gcsfs
python-decouple

# must match deployed prefect version
prefect[google,kubernetes]==1.2.2
